requirements.txt
annotated-types==0.7.0
anyio==4.10.0
boto3==1.40.27
botocore==1.40.27
certifi==2025.8.3
charset-normalizer==3.4.3
customtkinter==5.2.2
darkdetect==0.8.0
distro==1.9.0
ecdsa==0.19.1
envs==1.4
exceptiongroup==1.3.0
future==0.18.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==0.3.75
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.6
langsmith==0.4.27
openai==1.107.0
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
pycryptodome==3.3.1
pydantic==2.11.7
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose-cryptodome==1.3.2
PyYAML==6.0.2
PyJWT==2.10.1
requests==2.32.5
requests-toolbelt==1.0.0
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
warrant==0.6.1
xxhash==3.5.0
zstandard==0.24.0
